BF3431195 Webui Advanced Lookup not working See https://sourceforge.net/tracker/?func=detail&aid=3431195&group_id=176962&atid=955896
Update to FindWindow.java to fix the advanced lookup in ZK.  Also added
funtionality to remember the last search performed and a way to start a new
search.

Modified Find.java in the Swing client to have the same functionality and to
share searches.

Code review and testing required.  Some TODOs remain regarding translation of
text.