#1 Commit merging from Adaxa's to ADempiere repository, reference to issue: https://github.com/marcalwestf/adaxaFinReports/issues/1
First commit integrating Adaxa's Reports/Financial Reports into
ADempiere.
It was accomplished by merging manually classes from Adaxa's repository,
kindly supplied by Adaxa.
Due to the fact that the merging was done from Mercurial to Git and this
had to be done by hand, a possible history of one class' commits will
not visible. 
Please refer to Adaxa's repository if you are interested on a specific
class commit history.
By merging, sometimes the Adaxa class was taken as starting point and
the merging was accomplished by adding the differences to ADempiere;
sometimes it was done the other way: the ADempiere class was taken and
the differences to Adaxa were added.
Because of the huge differences between both repositories, some classed
were only partially merged, leaving it to the developers to catch up
with it.